llm-learning/labs/01-simple-llm-call/01-gemini-llm/requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0

